% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.package.R
\docType{package}
\name{cor.package-package}
\alias{cor.package}
\alias{cor.package-package}
\title{cor.package: A Package for Correlation Analysis}
\description{
This package provides functions to calculate correlation efficiently.
}
\author{
\strong{Maintainer}: Tong Liu \email{liutong.csu@gmail.com}

}
\keyword{internal}
