% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.package.R
\docType{package}
\name{cor.package-package}
\alias{cor.package}
\alias{cor.package-package}
\title{cor.package: A Package for Correlation Analysis}
\description{
The cor.package provides efficient and flexible tools for:
\itemize{
\item Calculating correlation coefficients using customizable methods (cor_new).
\item Handling missing values in datasets (handle_na).
\item Computing and visualizing correlation matrices for high-dimensional data (efficient_corr_matrix).
\item Calculating correlation coefficients more efficiently using Rcpp (cor_new_rcpp_wrapper).
}
\subsection{Main Functions}{
\itemize{
\item \code{cor_new()}: Computes the correlation between two numeric vectors.
\item \code{handle_na()}: Handles missing values in numeric datasets.
\item \code{efficient_corr_matrix()}: Computes the correlation at the high dimensional level.
\item \code{cor_new_rcpp_wrapper()}: Realizing cor_new() using C++.
}
}
}
\examples{
library(cor.package)
x = c(1, 2, 3, NA, 5)
y = c(2, 4, NA, 8, 10)
#Pearson correlation with mean imputation
cor_new(x, y, method = "pearson", na_handle = "mean")
#Spearman correlation with complete case handling
cor_new(x, y, method = "spearman", na_handle = "complete")
#Kendall correlation
cor_new(x, y, method = "kendall")


}
\author{
\strong{Maintainer}: Tong Liu \email{liutong.csu@gmail.com}

}
\keyword{internal}
